Moved common methods/ds to a different file
